added code to load ServerUser

+++ src/sidequest/model/user.h
@@ -9,7 +9,7 @@ namespace Sidequest
 
 	class User {
 	public:
-		typedef unsigned long Id;
+		typedef int Id;
 
 		User();
 		~User();
